audit: fix VFS-016 — vfsRead implementations use wrong offset type
ufs_readFile, ufs_writeFile (ufs.c), ffs_read (ffs.c), readUbixFS (ubixfs.c),
and devfs_read (devfs.c) all declared the offset parameter as uInt32/uint32_t/
long instead of off_t (64-bit on i386).  The vfsRead function pointer in vfs.h
uses off_t; a narrower type misaligns the subsequent 'size' argument on the
stack, causing every read to receive a wrong size value.

Fix: change all four implementations and their header declarations to off_t.
Add <sys/types.h> where missing.

+++ sys/include/fs/ufs/ffs.h
@@ -34,7 +34,7 @@
 #include <fs/vfs/vfs.h>
 #include <sys/bus.h>
 
-int ffs_read(fileDescriptor_t *,char *,uint32_t,long);
+int ffs_read(fileDescriptor_t *,char *,off_t,long);
 
 #endif
